requirements.txt
pip==23.0.1
rpi-gpio==0.7.1
setuptools==66.1.1
